FARMART
My description: while you press register button it automatically send an email to verify your account  , with meeting the validation as expected. Further more if you forget password it will send a resetting code through verifying your authenticity.

Expected errors there might be some fluctuations , like one page pertinently function and some times out of logic and  it might not for connectivity issues, so far I perceived this.

Expected feature plan to add multi-user login: had tried to work on it, but had  conflict of codes of no clue. We struggled to oversee the cause but unlikely we don't  so Will discover the secret bugs and come up with a good version soon.
Expected errors there might be some fluctuations it might not for connectivity issues, so far I perceived this.

Expected feature plan to add multi-user login: had tried to work on it, but had  conflict of codes of no clue. We struggled to oversee the cause but unlikely we don't  so Will discover the secret bugs and come up with a good version soon.